Large commit. Upgrade rails and re-do some of the features.
Remove all admin UI administration. Remove projects, site, and pages concepts completely. Reduce routes to 'show' routes
The idea with this is that if this rails app is just a blog only, it
will be easier to maintain and upgrade, it will have less dependencies
etc.

Rewrite the routes file, rails 3.x routes format changed considerably so
it was easier to rewrite them.

Rewrite the configuration files. It made sense to move some
configuration that was previously stored in the database via a "Site"
model to YAML configuration files.

Remove shoulda macros and contexts in favor of test unit. I prefer plain
test unit and no longer use cucumber or shoulda.

update to more modern way of specifying ruby version and gemset. Update to ruby 1.9.x version.

Fix the script directory. The scripts were moved from rails 2.x in favor
of the single "rails" script. Copied from another rails 3 project.

Add the rack config file. New requirement of rails 3.x.

Add a secret token. Rails 3.x requirement.

Remove the slugs plugin in favor of using a to_param implementation on
articles. The slugs are nicer URLs but the to_param style is simple and
makes upgrades easier.

Replace the tagging plugin (acts_as_taggable_on_steroids) with a simple
implementation for articles. The plugin was nice but it needs to be
converted to a gem. A couple of gem versions of the plugin did not work
correctly. To implement simple tagging just for the Article model was
not much code so this should be easier to maintain and make upgrades
easier since it does not require a dependency.

Implement comment author edit and delete
via cookies and session (which uses cookie store)

Add scripts to dump articles to yml files in the
articles directory, and a script to do the reverse
and take a yml file and insert it into the database.

Replace will_paginate with kaminari

Add redirect from old URL format to the new URL format

Add production config to website info yaml file

Add /contact route so that the old route is preserved

Make sure to supply the id when inserting article
records so the article_id foreign_key on comments
keeps working right