Convert dump and publish scripts to single script
Add generate task to make a new starter text file

Use the Jekyll style with a yaml chunk in the 
beginning of the file, then plain markdown text 
after that. Use a regex to parse out the yaml 
when publishing.

Taggings and articles can be deleted locally and
re-published, but need to smooth this out a bit more
yet so the same article can be modified locally
and edited, published, previewed, without creating
duplicate taggins.